Overview undetected synthetic miRNAs
Detection rate sensitivity for the synthetic miRNAs that could not be detected in at least one of the replicates for one of the kits and in at least one mix. The remaining miRNAs not represented in the plot were detected in all replicates of all kits. The library preparation kit and the replicate are presented on the y-axis. A darkened box within the plot indicates that the miRNA for this kit and replicate was not detected with at least one count.